+++ test/fixtures/automatic-labelling-folder/main.tf
@@ -18,6 +18,6 @@ module "automatic_labelling_folder" {
   source = "../../../examples/automatic-labelling-folder"
 
   project_id = var.project_id
-  folder_id  = var.folder_id
+  folder_id  = var.sub_folder_id
   region     = var.region
 }