+++ test/fixtures/dynamic-files/variables.tf
@@ -19,11 +19,6 @@ variable "project_id" {
   description = "The ID of the project to which resources will be applied."
 }
 
-variable "sub_folder_id" {
-  type        = string
-  description = "The ID of the folder to look for changes."
-}
-
 variable "region" {
   type        = string
   description = "The region in which resources will be applied."